requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
SQLAlchemy>=2.0,<3
Flask-Login>=0.6,<1
python-dotenv>=1.0,<2
Flask-Cors>=4.0,<5
pandas>=2.2,<3
click>=8.1,<9
